% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{coef.blbglm}
\alias{coef.blbglm}
\title{Compute the Bootstrap Coefficients}
\usage{
\method{coef}{blbglm}(fit)
}
\arguments{
\item{fit}{a vector containing the logistic regression models for each bootstrap sample.}
}
\value{
a list of bootstrap coefficient estimates calculated from the BLB estimates
}
\description{
\code{coef.blbglm} is used to find the coefficient estimates from the bag of little bootstraps
 by dividing the sum of each bootstrap sample coefficients by the number of samples.
}
