% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{predict.blbglm}
\alias{predict.blbglm}
\title{Apply logistic regression to Bags of Little Bootstraps (BLB)}
\usage{
\method{predict}{blbglm}(fit, testdata, confidence = FALSE)
}
\arguments{
\item{fit}{a list of fitted models computed using BLB (e.g. fit<-blbglm(...)).}

\item{testdata}{a tibble, data frame, or named vector containing the data.}

\item{confidence}{logical. If TRUE, a confidence interval will be returned.}
}
\description{
\code{predict.blbglm} predicts the probability of obtaining a response variable, and returns
a 95% confidence interval for the probability if confidence = TRUE.
}
\examples{
fit<-blbglm(income~age+`hours-per-week`,data=adult)
testdata<-tibble(age=c(25,30), `hours-per-week`=c(50,100))
predict.blbglm(fit,testdata,confidence=T)
             [,1]           [,2]
2.5\%     0.1863978     0.09907537
97.5\%   0.1960761      0.10568005
}
