% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{sigma.blbglm}
\alias{sigma.blbglm}
\title{Compute Sigma and Its Confidence Interval for Each Bootstrap Model}
\usage{
\method{sigma}{blbglm}(fit, ci = FALSE)
}
\arguments{
\item{fit}{a list of models for the bootstrap samples.}

\item{ci}{logical. If TRUE, the function will return the 95 percent confidence intervals for each bootstrap model.}
}
\description{
\code{sigma.blbglm} is used to compute the overall sigma and its 95 percent CI for the whole dataset.
}
