% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\docType{package}
\name{STAproject-package}
\alias{STAproject}
\alias{STAproject-package}
\title{STAproject: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\details{
Logistic Regression with Little Bag of Bootstraps
}
