% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{blbcoef}
\alias{blbcoef}
\title{Compute the coefficients}
\usage{
blbcoef(fit)
}
\arguments{
\item{fit}{logistic regression model for each subsample.}
}
\description{
Compute Coefficients from Each Model
}
