% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{confint.blbglm}
\alias{confint.blbglm}
\title{Compute Confidence Interval for Each BLB Coefficient}
\usage{
\method{confint}{blbglm}(fit)
}
\arguments{
\item{fit}{a list of bootstrap models for the BLB subsamples.}
}
\description{
\code{confint.blbglm} is used to compute the 95 percent confidence interval
 for each coefficient using Bag of Little Bootstraps.
}
\examples{
confint.blbglm(blbfit)
##       (Intercept) hours.per.week        age genderMale
## 2.5\%   -11.895876     0.04655235 0.01985707  0.4567024
## 97.5\%   -4.948412     0.15050993 0.07731212  2.8073899
}
