% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{glm1}
\alias{glm1}
\title{logistic regression estimates}
\usage{
glm1(formula, data, freqs)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class)}

\item{data}{the sub_data from original data set}

\item{freqs}{the frequence from glm_each_boot function}
}
\description{
estimate the regression estimates based on given number of repetitions
}
