% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{glm1}
\alias{glm1}
\title{Logistic regression estimates}
\usage{
glm1(formula, data, freqs)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class).}

\item{data}{the sub_data from original data set.}

\item{freqs}{the frequency from glm_each_boot function.}
}
\description{
Estimate Regression Estimates Based on a Given Number of Repetitions.
}
\examples{
form = income ~ hours.per.week + age + gender
freq <- rmultinom(1, nrow(vignette_sample), rep(1, nrow(vignette_sample)))
glm1(form, vignette_sample, freq)
## $sigma
## [1] 1.17456
##
## $coef
##    (Intercept) hours.per.week            age     genderMale
##    -4.94910851     0.05414013     0.02121227     0.64940167
}
