% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{glm_each_subsample}
\alias{glm_each_subsample}
\title{Compute the estimates}
\usage{
glm_each_subsample(formula, data, n, B)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class).}

\item{data}{the sub_data from original data set.}

\item{n}{the size of original data set.}

\item{B}{the subsample repeat B times.}
}
\description{
Compute the Estimates for Each Subsample and Fit mMdel for Each Subsmple B Times.
}
\examples{
form = income ~ hours.per.week + age + gender
glm_each_subsample(form, vignette_sample, nrow(df), B=100)
}
